Modularize API and code review
- Break the classes contained with the large ProbeInterface class into
  standalone files
- Remove autogenated backing fields and cruft
- Some sytanx changes and simplifications
